Update index.html
Got Moves to show what level they are learned at and how they are learned. If the move is learned via Egg, Tutor, or TM. If it has the method marked as Egg, Tutor or TM it will remove the 0 for the level and add (TM), (Tutor), or (Egg) depending on the move.